change ktap_number to "long"
Since we are going to use 64 bit data in script,
repesent ip value or memory pointer, so "int" as
number is not enought for 64 bit data operation,
change ktap_number to "long" now.

There would not bring any performance issue.

There have one issue about read wrong 64 bit value
in compiler, use strtol to fix the issue.

All testcase are passed with this change.

The origin issue was reported by Qingping.

Reported-by: Qingping Hou <qingping.hou@gmail.com>

+++ include/ktap_types.h
@@ -389,7 +389,7 @@ union ktap_gcobject {
 #define check_exp(c,e)                (e)
 
 
-typedef int ktap_number;
+typedef long ktap_number;
 
 
 #define ktap_number2int(i,n)   ((i)=(int)(n))
